Add precompilation
Especially on (the upcoming) Julia 1.9, this greatly reduces TTFX,
from >5s to ~10ms on my machine.

Because SnoopPrecompile does not support Julia 0.7, this PR requires
that we abandon support for very old Julia versions, and that in turn
requires a minor version bump (policy of the General registry).
Since this package isn't changing quickly, I recommend taking the leap
to 1.0.