Fix failing test (#69)
The greedycd algorithm gets tested for its ability to reconstruct
ground-truth from offset initialization. However, it also tests
the effects of regularization (sparseness priors), and with
regularization the "naive ground truth" is not the actual ground truth
anymore. This reduces the amplitude of the regularization coefficient
to reduce the perturbation and allow the test to pass.